requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
flask-smorest==0.47.0
marshmallow==4.2.2
marshmallow-sqlalchemy==1.4.2
redis==5.0.1
rq==1.15.1
python-dotenv==0.21.0
psycopg2-binary==2.9.7